refactor: restructure folder tree view component and endpoints
