+++ src/components/styles.scss
@@ -13,7 +13,7 @@
 .treeitem {
     width: 100%;
     text-align: left;
-    background-color: white;
+    background-color: var(--theme-elevation-0);
     padding: 6px 10px;
     position: relative;
     border-radius: 8px;
@@ -22,46 +22,31 @@
     display: flex;
     align-items: center;
 }
-.treeitem:hover {
-    background-color: var(--selected-color);
-}
 
-.tree button[role="treeitem"]:focus {
-    outline: none;
+.treeitem:hover {
+    background-color: var(--theme-elevation-50);
 }
 
 .treeitem.selected {
-    background-color: #eee;
+    background-color: var(--theme-elevation-100);
 }
 
-button:focus-visible .treeitem.focused, .treeitem.searchmatch.focused {
-    outline: 2px solid black;
+.treeitem > .icon--chevron {
+    display: none;
 }
 
-.treeitem.drop {
-    border-color: var(--selected-color);
-    background-color: #e1f1f8;
-}
-
-.treeitem.searchmatch {
-    background-color: #e1f8ff;
-}
-
-.treeitem.folder:before {
+.treeitem.folder > .icon--chevron {
+    display: block;
     flex-shrink: 0;
-    content: url(data:image/svg+xml;base64,PHN2ZyB2ZXJzaW9uPSIxLjEiIHhtbG5zPSJodHRwOi8vd3d3LnczLm9yZy8yMDAwL3N2ZyIgeG1sbnM6eGxpbms9Imh0dHA6Ly93d3cudzMub3JnLzE5OTkveGxpbmsiIHg9IjBweCIgeT0iMHB4IiB2aWV3Qm94PSIwIDAgMTYgMTYiIGVuYWJsZS1iYWNrZ3JvdW5kPSJuZXcgMCAwIDE2IDE2IiB4bWw6c3BhY2U9InByZXNlcnZlIj48Zz48Zz48cGF0aCBmaWxsLXJ1bGU9ImV2ZW5vZGQiIGNsaXAtcnVsZT0iZXZlbm9kZCIgZD0iTTQuNjQ2IDEuNjQ2YS41LjUgMCAwIDEgLjcwOCAwbDYgNmEuNS41IDAgMCAxIDAgLjcwOGwtNiA2YS41LjUgMCAwIDEtLjcwOC0uNzA4TDEwLjI5MyA4IDQuNjQ2IDIuMzU0YS41LjUgMCAwIDEgMC0uNzA4eiIgY2xhc3M9InJjdC10cmVlLWl0ZW0tYXJyb3ctcGF0aCI+PC9wYXRoPjwvZz48L2c+PC9zdmc+);
-    width: 10px;
-    display: inline-block;
-    z-index: 1;
-    margin-right: 4px;
     transition: transform 0.1s ease-in-out;
+    transform: rotate(-90deg);
 }
 
-.treeitem.folder.expanded:before {
-    transform: rotate(90deg);
+.treeitem.folder.expanded > .icon--chevron {
+    transform: rotate(0deg);
 }
 
 .treeitem:not(.folder) {
-    padding-left: 24px;
+    padding-left: 16px;
 }
