+++ src/types.ts
@@ -1,12 +1,15 @@
 import type { Endpoint } from "payload";
 
 export interface Endpoints {
-  files: Omit<Endpoint, 'root'>;
-  openFolder: Omit<Endpoint, 'root'>;
+  folder: Omit<Endpoint, 'root'>;
+  item: Omit<Endpoint, 'root'>;
+  root: Omit<Endpoint, 'root'>;
 }
 
-export type File = {
+export type TreeData = {
+  createdAt: string;
+  data?: TreeData[];
   id: string;
-  relationTo: string;
   title: string;
+  updatedAt: string;
 }
